# backend/src/modules/profile/routes.ts
/**
 * Profile Routes
 * 
 * Implements profile management endpoints per contracts/profiles.openapi.yml
 * 
 * Tasks: T071-T080
 * 
 * Endpoints:
 * - POST /api/profile (T072) - Create profile from questionnaire
 * - GET /api/profile (T075) - Retrieve current user profile
 * - PATCH /api/profile (T076) - Update existing profile
 * - GET /api/profile/export (T079) - Data portability (JSON export)
 * - DELETE /api/profile/delete (T080) - Soft delete with 30-day fulfillment
 */

import { Router } from 'express';

const router = Router();

// TODO: Task T072 - POST /api/profile endpoint
router.post('/', async (req, res) => {
  // Implementation: M1-M2 - Week 3-4
  // - Validate profile input with Zod schema (T073)
  // - Store profile with encrypted sensitive fields (T074)
  // - Create profile from questionnaire data
  res.status(501).json({ error: 'Not Implemented' });
});

// TODO: Task T075 - GET /api/profile endpoint
router.get('/', async (req, res) => {
  // Implementation: M1-M2 - Week 3-4
  // - Retrieve current user profile
  // - Return profile data
  res.status(501).json({ error: 'Not Implemented' });
});

// TODO: Task T076 - PATCH /api/profile endpoint
router.patch('/', async (req, res) => {
  // Implementation: M1-M2 - Week 3-4
  // - Update existing profile
  // - Handle partial updates (T077)
  // - Invalidate recommendation cache (T078)
  res.status(501).json({ error: 'Not Implemented' });
});

// TODO: Task T079 - GET /api/profile/export endpoint
router.get('/export', async (req, res) => {
  // Implementation: M5 - Week 9 (Privacy & Data Rights)
  // - Generate JSON export
  // - Include profile data, analytics (if opted in), recommendation history
  res.status(501).json({ error: 'Not Implemented' });
});

// TODO: Task T080 - DELETE /api/profile/delete endpoint
router.delete('/delete', async (req, res) => {
  // Implementation: M5 - Week 9 (Privacy & Data Rights)
  // - Soft delete
  // - Schedule purge after 30 days
  res.status(501).json({ error: 'Not Implemented' });
});

export default router;